ボタン画像の色変更 part2

+++ fixIconAlpha.pl
@@ -29,52 +29,69 @@ sub fix($){
 	my $width = $image->width;
 	my $height = $image->height;
 	my $alphaMax = 0;
-	for( my $y = 0; $y < $height ;++$y){
+	my $isColor = 0 ;
+Loop: for( my $y = 0; $y < $height ;++$y){
 		for( my $x = 0; $x < $width ;++$x){
 			my $c = $image->getPixel($x,$y);
 			my $c0 = ($c >> 24)&255; # GDのalphaは0が不透明、127が完全透明
+			my $r = ($c >> 16)&255; 
+			my $g = ($c >> 8)&255; 
+			my $b = ($c >> 0)&255; 
+			if( abs($r-$g) >= 2 or abs($r-$b) >= 2 or abs($g-$b) >= 2 ){
+				$isColor = 1;
+				last Loop;
+			}
 			my $alpha= 127-$c0;
 			if( $alpha > $alphaMax){
 				$alphaMax = $alpha;
 			}
 		}
 	}
-
-	return if $alphaMax == 127;
+	
+	return if $isColor;
+	return if $alphaMax == 0;
 
 	say "$width x $height ,alphaMax=$alphaMax $relpath";
 
+	if( not -f $dstfile ){
 
-	my $scale = 127.0/$alphaMax;
+		my $scale = 127.0/$alphaMax;
 
-	for( my $y = 0; $y < $height ;++$y){
-		for( my $x = 0; $x < $width ;++$x){
-			my $c = $image->getPixel($x,$y);
-			my $c0 = ($c >> 24)&255; # GDのalphaは0が不透明、127が完全透明
-			next if $c0 == 127;
-			my $alpha= 127-$c0;
-			$c0 = 127 - int( 0.5 + $alpha * $scale );
-			$c = ($c & 0xffffff ) | ($c0 << 24 );
-			$image->setPixel($x,$y,$c);
+		for( my $y = 0; $y < $height ;++$y){
+			for( my $x = 0; $x < $width ;++$x){
+				my $c = $image->getPixel($x,$y);
+				my $c0 = ($c >> 24)&255; # GDのalphaは0が不透明、127が完全透明
+				my $r = ($c >> 16)&255; 
+				my $g = ($c >> 8)&255; 
+				my $b = ($c >> 0)&255; 
+				next if $c0 == 127;
+				my $alpha= 127-$c0;
+				$c0 = 127 - int( 0.5 + $alpha * $scale );
+				my($rgb) = ($r+$g+$b >= 555 ? 0xffffff : 0x000000 );
+				$c = $rgb | ($c0 << 24 );
+				$image->setPixel($x,$y,$c);
+			}
 		}
-	}
-	
-	my $dstdir = $dstfile;
-	$dstdir =~ s|/[^/]+\z||;
-	make_path($dstdir);
+		
+		my $dstdir = $dstfile;
+		$dstdir =~ s|/[^/]+\z||;
+		make_path($dstdir);
 
-	$image->saveAlpha(1);
-	open(my $fh,">",$dstfile) or die "$dstfile $!";
-	binmode $fh;
-	print $fh $image->png;
-	close($fh) or die "$dstfile $!";
-
-	my $relpath_webp = $relpath;
-	$relpath_webp =~ s/(png|webp)\z/webp/i;
-	if( "png" eq lc $1 ){
-		unlink $relpath;
+		$image->saveAlpha(1);
+		open(my $fh,">",$dstfile) or die "$dstfile $!";
+		binmode $fh;
+		print $fh $image->png;
+		close($fh) or die "$dstfile $!";
 	}
-	system qq(cwebp $dstfile -quiet -o $relpath_webp);
+
+	if(1){
+		my $relpath_webp = $relpath;
+		$relpath_webp =~ s/(png|webp)\z/webp/i;
+		if( "png" eq lc $1 ){
+			unlink $relpath;
+		}
+		system qq(cwebp $dstfile -quiet -o $relpath_webp);
+	}
 }
 
 
@@ -86,7 +103,7 @@ find(
 			my $relpath = $File::Find::name;
 			return if -d $relpath;
 			return if not $relpath =~ m|/drawable-|;
-			return if $relpath =~ m|/media_type_|;
+			return if $relpath =~ m%/(media_type_|emoji_)%;
 			push @files,$relpath;
 		}
 	}
